Convert RequestTypeChooser to Typescript
